Move yxml into shlr
