+++ libr/bin/format/xnu/r_cf_dict.c
@@ -3,7 +3,7 @@
 #include <r_util.h>
 #include <r_list.h>
 
-#include "yxml.h"
+#include "../../../shlr/yxml/yxml.h"
 #include "r_cf_dict.h"
 
 #define XMLBUFSIZE 4096
