+++ libr/bin/meson.build
@@ -88,7 +88,6 @@ r_bin_sources = [
   'format/mach0/mach0.c',
   'format/mach0/mach064.c',
   'format/xnu/r_cf_dict.c',
-  'format/xnu/yxml.c',
   'format/mdmp/mdmp.c',
   'format/mdmp/mdmp_pe.c',
   'format/mdmp/mdmp_pe64.c',
@@ -144,7 +143,8 @@ r_bin = library('r_bin', r_bin_sources,
     r_socket_dep,
     r_syscall_dep,
     java_dep,
-    lz4_dep
+    lz4_dep,
+    yxml_dep
   ],
   install: true,
   implicit_include_directories: false,
